Update import paths after section-templates restructure (#1115)
