+++ projects/editor/modules/section-templates/components/stimulus/audio1-stimulus/audio1-stimulus.component.ts
@@ -1,4 +1,4 @@
-import { Audio1StimulusOptions } from 'editor/modules/section-templates/stimulus-interfaces';
+import { Audio1StimulusOptions } from 'editor/modules/section-templates/models/stimulus-interfaces';
 import { Component, EventEmitter, Output } from '@angular/core';
 import { FileService } from 'common/services/file.service';
