+++ projects/editor/modules/section-templates/components/stimulus/message-stimulus/message-stimulus.component.ts
@@ -1,4 +1,4 @@
-import { MessageStimulusOptions } from 'editor/modules/section-templates/stimulus-interfaces';
+import { MessageStimulusOptions } from 'editor/modules/section-templates/models/stimulus-interfaces';
 import { Component } from '@angular/core';
 
 @Component({
