+++ projects/editor/modules/section-templates/services/template.service.ts
@@ -3,7 +3,7 @@ import { MatDialog } from '@angular/material/dialog';
 import { TranslateService } from '@ngx-translate/core';
 import {
   RadioWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/radio/radio-dialog/radio-dialog.component';
+} from 'editor/modules/section-templates/components/radio/radio-dialog/radio-dialog.component';
 import { ElementFactory } from 'common/utils/element-factory';
 import { PositionProperties, PropertyGroupGenerators } from 'common/models/elements/property-group-interfaces';
 import { UnitService } from 'editor/src/app/services/unit.service';
@@ -11,53 +11,53 @@ import { IDService } from 'editor/src/app/services/id.service';
 import { UIElement } from 'common/models/elements/element';
 import {
   LikertWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/likert-dialog/likert-dialog.component';
+} from 'editor/modules/section-templates/components/likert-dialog/likert-dialog.component';
 import {
   InputWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/text-input-dialog/text-input-dialog.component';
+} from 'editor/modules/section-templates/components/text-input-dialog/text-input-dialog.component';
 import {
   MarkingPanelDialogComponent
-} from 'editor/modules/section-templates/dialogs/marking-panel-dialog/marking-panel-dialog.component';
+} from 'editor/modules/section-templates/components/marking-panel-dialog/marking-panel-dialog.component';
 import {
   GeometryWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/geometry-dialog/geometry-dialog.component';
+} from 'editor/modules/section-templates/components/geometry-dialog/geometry-dialog.component';
 import {
   DroplistWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/droplist-dialog/droplist-dialog.component';
+} from 'editor/modules/section-templates/components/droplist-dialog/droplist-dialog.component';
 import {
   MathTableWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/mathtable-dialog/mathtable-dialog.component';
+} from 'editor/modules/section-templates/components/mathtable-dialog/mathtable-dialog.component';
 import {
   Text3WizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/text3-dialog/text3-dialog.component';
+} from 'editor/modules/section-templates/components/text3-dialog/text3-dialog.component';
 import {
   CheckboxWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/checkbox-dialog/checkbox-dialog.component';
+} from 'editor/modules/section-templates/components/checkbox-dialog/checkbox-dialog.component';
 import { SelectionService } from 'editor/src/app/services/selection.service';
 import { PositionedUIElement, UIElementType } from 'common/models/ui-element-interfaces';
 import { TextImageLabel } from 'common/models/label-interfaces';
 import { PageProperties } from 'common/models/page';
-import * as TextBuilders from 'editor/modules/section-templates/builders/text-builders';
-import * as TextInputBuilders from 'editor/modules/section-templates/builders/text-input-builders';
-import * as RadioBuilders from 'editor/modules/section-templates/builders/radio-builders';
-import * as CheckboxBuilders from 'editor/modules/section-templates/builders/checkbox-builders';
-import * as DroplistBuilders from 'editor/modules/section-templates/builders/droplist-builders';
-import * as GeometryBuilders from 'editor/modules/section-templates/builders/geometry-builders';
-import * as MathtableBuilders from 'editor/modules/section-templates/builders/mathtable-builders';
-import * as StimulusBuilders from 'editor/modules/section-templates/builders/stimulus/stimulus-builders';
+import * as TextBuilders from 'editor/modules/section-templates/utils/text-builders';
+import * as TextInputBuilders from 'editor/modules/section-templates/utils/text-input-builders';
+import * as RadioBuilders from 'editor/modules/section-templates/utils/radio-builders';
+import * as CheckboxBuilders from 'editor/modules/section-templates/utils/checkbox-builders';
+import * as DroplistBuilders from 'editor/modules/section-templates/utils/droplist-builders';
+import * as GeometryBuilders from 'editor/modules/section-templates/utils/geometry-builders';
+import * as MathtableBuilders from 'editor/modules/section-templates/utils/mathtable-builders';
+import * as StimulusBuilders from 'editor/modules/section-templates/utils/stimulus/stimulus-builders';
 import {
   ClassicTemplateOptions, SortTemplateOptions, TwoPageTemplateOptions
-} from 'editor/modules/section-templates/droplist-interfaces';
+} from 'editor/modules/section-templates/models/droplist-interfaces';
 import {
   StimulusWizardDialogComponent
-} from 'editor/modules/section-templates/dialogs/stimulus/stimulus-dialog/stimulus-dialog.component';
+} from 'editor/modules/section-templates/components/stimulus/stimulus-dialog/stimulus-dialog.component';
 import {
   Audio1StimulusOptions, Audio2StimulusOptions,
   EmailStimulusOptions,
   MessageStimulusOptions,
   TextStimulusOptions
-} from 'editor/modules/section-templates/stimulus-interfaces';
-import { ImageRadioOptions, TextRadioOptions } from 'editor/modules/section-templates/radio-interfaces';
+} from 'editor/modules/section-templates/models/stimulus-interfaces';
+import { ImageRadioOptions, TextRadioOptions } from 'editor/modules/section-templates/models/radio-interfaces';
 import { TextElement } from 'common/models/elements/text-group-elements/text';
 import { EditorPage } from 'editor/src/app/models/editor-page';
 import { EditorSection } from 'editor/src/app/models/editor-section';
