+++ projects/editor/modules/section-templates/utils/mathtable-builders.ts
@@ -1,6 +1,6 @@
 import { PositionedUIElement } from 'common/models/ui-element-interfaces';
 import { SectionProperties } from 'common/models/section';
-import { TemplateService } from 'editor/modules/section-templates/template.service';
+import { TemplateService } from 'editor/modules/section-templates/services/template.service';
 import { IDService } from 'editor/src/app/services/id.service';
 import { CONSTANTS } from 'editor/modules/section-templates/constants';
 import { EditorSection } from 'editor/src/app/models/editor-section';
