+++ projects/editor/modules/section-templates/utils/radio-builders.ts
@@ -1,7 +1,7 @@
 import { TextImageLabel } from 'common/models/label-interfaces';
-import { TemplateService } from 'editor/modules/section-templates/template.service';
+import { TemplateService } from 'editor/modules/section-templates/services/template.service';
 import { IDService } from 'editor/src/app/services/id.service';
-import { ImageRadioOptions, TextRadioOptions } from 'editor/modules/section-templates/radio-interfaces';
+import { ImageRadioOptions, TextRadioOptions } from 'editor/modules/section-templates/models/radio-interfaces';
 import { EditorSection } from 'editor/src/app/models/editor-section';
 
 export function createTextRadioSection(options: TextRadioOptions, idService: IDService): EditorSection {
