+++ projects/editor/modules/section-templates/utils/stimulus/email.ts
@@ -1,4 +1,4 @@
-import { EmailStimulusOptions } from 'editor/modules/section-templates/stimulus-interfaces';
+import { EmailStimulusOptions } from 'editor/modules/section-templates/models/stimulus-interfaces';
 
 export function getEmailTemplateString(
   options: EmailStimulusOptions & { fromLabel: string, toLabel: string, subjectLabel: string, sendLabel: string }) {
