+++ projects/editor/modules/section-templates/utils/stimulus/stimulus-builders.spec.ts
@@ -1,6 +1,6 @@
 import { IDService } from 'editor/src/app/services/id.service';
-import { EmailStimulusOptions } from 'editor/modules/section-templates/stimulus-interfaces';
-import { createEmailSection } from 'editor/modules/section-templates/builders/stimulus/stimulus-builders';
+import { EmailStimulusOptions } from 'editor/modules/section-templates/models/stimulus-interfaces';
+import { createEmailSection } from 'editor/modules/section-templates/utils/stimulus/stimulus-builders';
 
 describe('createEmailSection', () => {
   let idService: IDService;
