+++ projects/editor/modules/section-templates/utils/text-builders.ts
@@ -1,7 +1,7 @@
 import { SectionProperties } from 'common/models/section';
 import { PositionedUIElement } from 'common/models/ui-element-interfaces';
 import { IDService } from 'editor/src/app/services/id.service';
-import { TemplateService } from 'editor/modules/section-templates/template.service';
+import { TemplateService } from 'editor/modules/section-templates/services/template.service';
 import { EditorSection } from 'editor/src/app/models/editor-section';
 
 // Disable linting rules to have smaller code
